API work and documentation

+++ lua/simctl/api/listapps.lua
@@ -2,6 +2,9 @@ local M = {}
 
 local util = require("simctl.lib.util")
 local simctl = require("simctl.lib.simctl")
+local aw = require("simctl.lib.async")
+local pickers = require("simctl.lib.pickers")
+local config = require("simctl.config")
 
 local function filterAppsByType(parentTbl, appType)
 	local filteredApps = {}
@@ -18,52 +21,71 @@ M.AppType = {
 	System = "System",
 }
 
+--- List installed apps
+-- @param args.appType string Optional type of app to filter by, User or System
+-- @param callback function indicating success or failure and a table of apps
 M.listapps = function(args, callback)
 	callback = callback or function() end
+	args = args or {}
 
-	args = util.merge(args or {}, {
-		deviceId = "booted",
-	})
-
-	simctl.execute({ "listapps", "booted" }, function(return_val, humane, stdout, stderr)
-		if return_val ~= 0 then
-			local message = humane or stderr
-			util.notify(message)
-
-			callback(false, nil, stdout, stderr)
-			return
+	aw.async(function()
+		if args.deviceId == nil and config.options.devicePicker then
+			args.deviceId = aw.await(pickers.pickDevice)
 		end
 
-		vim.schedule(function()
-			local tmpFilePath = os.tmpname()
-			local file = io.open(tmpFilePath, "w")
-			if file then
-				file:write(stdout)
-				file:close()
-			else
-				callback(false)
+		args = util.merge(args, {
+			deviceId = "booted",
+		})
+
+		simctl.execute({ "listapps", args.deviceId }, function(return_val, humane, stdout, stderr)
+			if return_val ~= 0 then
+				local message = humane or stderr
+				util.notify(message)
+
+				callback(false, nil, stdout, stderr)
 				return
 			end
 
-			local command = string.format('plutil -convert json -o - "%s"', tmpFilePath)
-			local jsonOutput = vim.fn.system(command)
+			vim.schedule(function()
+				local tmpFilePath = os.tmpname()
+				local file = io.open(tmpFilePath, "w")
+				if file then
+					file:write(stdout)
+					file:close()
+				else
+					callback(false)
+					return
+				end
 
-			-- Check for error
-			if vim.v.shell_error ~= 0 then
-				callback(false, nil, jsonOutput, nil)
+				local command = string.format('plutil -convert json -o - "%s"', tmpFilePath)
+				local jsonOutput = vim.fn.system(command)
+
+				if vim.v.shell_error ~= 0 then
+					callback(false, nil, jsonOutput, nil)
+					os.remove(tmpFilePath)
+					return
+				end
+
+				local keyValueApps = vim.json.decode(jsonOutput)
+				if not keyValueApps then
+					callback(false, nil, jsonOutput, nil)
+					os.remove(tmpFilePath)
+					return
+				end
+
+				local apps = {}
+				for _, value in pairs(keyValueApps) do
+					table.insert(apps, value)
+				end
+
+				if args.appType then
+					apps = filterAppsByType(apps, args.appType)
+				end
+
 				os.remove(tmpFilePath)
-				return
-			end
 
-			local apps = vim.json.decode(jsonOutput)
-
-			if args.appType then
-				apps = filterAppsByType(apps, args.appType)
-			end
-
-			os.remove(tmpFilePath)
-
-			callback(true, apps)
+				callback(true, apps)
+			end)
 		end)
 	end)
 end
